proyecto_integrado/backend/requirements.txt
django>=4.2,<5.0
pillow>=10.0
psycopg2-binary>=2.9
python-decouple>=3.8
whitenoise>=6.5
gunicorn>=21.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
